+++ src/test/groovy/ru/vyarus/guice/persist/orient/repository/core/result/projection/RepositoryResultProjections.groovy
@@ -21,8 +21,7 @@ public interface RepositoryResultProjections {
     @Query("select count(@rid) from Model")
     int getCount()
 
-    // automatic projection will not work here, but due to type erasure this will work even in runtime,
-    // except that actual result would be List<ODocument>
+    // orient will return list of ODocument and result will be "unwrapped" to raw values
     @Query("select name from Model")
     List<String> getNames()
 
@@ -40,7 +39,6 @@ public interface RepositoryResultProjections {
     @Query(value = "select name from VertexModel", connection = DbType.GRAPH)
     String[] getGraphNamesArray()
 
-    // in runtime result projection will not occur List<Vertex> will return (it's valid due to type erasure, but confusing)
     @Query(value = "select name from VertexModel", connection = DbType.GRAPH)
     List<String> getGraphNames()